Recategorize several packager unit tests as functional
There were packager tests in the unit/ test
directory that really were functional tests because they:

* Touched the filesystem
* Executed subprocesses

This was becomming a hindrance to quick feedback cycles
when running the unit test suite, and notable the unit/deploy/
suite.

Before:

$ time py.test tests/unit/
...
434 passed, 1 skipped in 6.70 seconds

$ time py.test tests/unit/deploy/
122 passed in 3.69 seconds

After:

$ time py.test tests/unit
423 passed, 1 skipped in 2.93 seconds

$ time py.test tests/unit/deploy/
111 passed in 0.46 seconds
