+++ tests/functional/test_package.py
@@ -1,9 +1,12 @@
 import os
 import zipfile
-import mock
+import tarfile
+import io
 from collections import defaultdict
 
 import pytest
+import mock
+
 from chalice.config import Config
 from chalice import Chalice
 from chalice import package
@@ -11,6 +14,9 @@ from chalice.deploy.packager import PipRunner
 from chalice.deploy.packager import DependencyBuilder
 from chalice.deploy.packager import Package
 from chalice.deploy.packager import MissingDependencyError
+from chalice.deploy.packager import SubprocessPip
+from chalice.deploy.packager import SDistMetadataFetcher
+from chalice.deploy.packager import InvalidSourceDistributionNameError
 from chalice.compat import lambda_abi
 from chalice.compat import pip_no_compile_c_env_vars
 from chalice.compat import pip_no_compile_c_shim
@@ -34,6 +40,11 @@ def sample_app():
         return {"hello": "world"}
 
     return app
+
+
+@pytest.fixture
+def sdist_reader():
+    return SDistMetadataFetcher()
 
 
 class PathArgumentEndingWith(object):
@@ -718,3 +729,159 @@ def test_will_create_outdir_if_needed(tmpdir):
     contents = os.listdir(str(outdir))
     assert 'deployment.zip' in contents
     assert 'sam.json' in contents
+
+
+class TestSubprocessPip(object):
+    def test_can_invoke_pip(self):
+        pip = SubprocessPip()
+        rc, err = pip.main(['--version'])
+        # Simple assertion that we can execute pip and it gives us some output
+        # and nothing on stderr.
+        assert rc == 0
+        assert err == b''
+
+    def test_does_error_code_propagate(self):
+        pip = SubprocessPip()
+        rc, err = pip.main(['badcommand'])
+        assert rc != 0
+        # Don't want to depend on a particular error message from pip since it
+        # may change if we pin a differnet version to Chalice at some point.
+        # But there should be a non-empty error message of some kind.
+        assert err != b''
+
+
+class TestSdistMetadataFetcher(object):
+    _SETUPTOOLS = 'from setuptools import setup'
+    _DISTUTILS = 'from distutils.core import setup'
+    _BOTH = (
+        'try:\n'
+        '    from setuptools import setup\n'
+        'except ImportError:\n'
+        '    from distutils.core import setuptools\n'
+    )
+
+    _SETUP_PY = (
+        '%s\n'
+        'setup(\n'
+        '    name="%s",\n'
+        '    version="%s"\n'
+        ')\n'
+    )
+
+    def _write_fake_sdist(self, setup_py, directory, ext):
+        filename = 'sdist.%s' % ext
+        path = '%s/%s' % (directory, filename)
+        if ext == 'zip':
+            with zipfile.ZipFile(path, 'w',
+                                 compression=zipfile.ZIP_DEFLATED) as z:
+                z.writestr('sdist/setup.py', setup_py)
+        else:
+            with tarfile.open(path, 'w:gz') as tar:
+                tarinfo = tarfile.TarInfo('sdist/setup.py')
+                tarinfo.size = len(setup_py)
+                tar.addfile(tarinfo, io.BytesIO(setup_py.encode()))
+        filepath = os.path.join(directory, filename)
+        return filepath
+
+    def test_setup_tar_gz(self, osutils, sdist_reader):
+        setup_py = self._SETUP_PY % (
+            self._SETUPTOOLS, 'foo', '1.0'
+        )
+        with osutils.tempdir() as tempdir:
+            filepath = self._write_fake_sdist(setup_py, tempdir, 'tar.gz')
+            name, version = sdist_reader.get_package_name_and_version(
+                filepath)
+        assert name == 'foo'
+        assert version == '1.0'
+
+    def test_setup_tar_gz_hyphens_in_name(self, osutils, sdist_reader):
+        # The whole reason we need to use the egg info to get the name and
+        # version is that we cannot deterministically parse that information
+        # from the filenames themselves. This test puts hyphens in the name
+        # and version which would break a simple ``split("-")`` attempt to get
+        # that information.
+        setup_py = self._SETUP_PY % (
+            self._SETUPTOOLS, 'foo-bar', '1.0-2b'
+        )
+        with osutils.tempdir() as tempdir:
+            filepath = self._write_fake_sdist(setup_py, tempdir, 'tar.gz')
+            name, version = sdist_reader.get_package_name_and_version(
+                filepath)
+        assert name == 'foo-bar'
+        assert version == '1.0-2b'
+
+    def test_setup_zip(self, osutils, sdist_reader):
+        setup_py = self._SETUP_PY % (
+            self._SETUPTOOLS, 'foo', '1.0'
+        )
+        with osutils.tempdir() as tempdir:
+            filepath = self._write_fake_sdist(setup_py, tempdir, 'zip')
+            name, version = sdist_reader.get_package_name_and_version(
+                filepath)
+        assert name == 'foo'
+        assert version == '1.0'
+
+    def test_distutil_tar_gz(self, osutils, sdist_reader):
+        setup_py = self._SETUP_PY % (
+            self._DISTUTILS, 'foo', '1.0'
+        )
+        with osutils.tempdir() as tempdir:
+            filepath = self._write_fake_sdist(setup_py, tempdir, 'tar.gz')
+            name, version = sdist_reader.get_package_name_and_version(
+                filepath)
+        assert name == 'foo'
+        assert version == '1.0'
+
+    def test_distutil_zip(self, osutils, sdist_reader):
+        setup_py = self._SETUP_PY % (
+            self._DISTUTILS, 'foo', '1.0'
+        )
+        with osutils.tempdir() as tempdir:
+            filepath = self._write_fake_sdist(setup_py, tempdir, 'zip')
+            name, version = sdist_reader.get_package_name_and_version(
+                filepath)
+        assert name == 'foo'
+        assert version == '1.0'
+
+    def test_both_tar_gz(self, osutils, sdist_reader):
+        setup_py = self._SETUP_PY % (
+            self._BOTH, 'foo-bar', '1.0-2b'
+        )
+        with osutils.tempdir() as tempdir:
+            filepath = self._write_fake_sdist(setup_py, tempdir, 'tar.gz')
+            name, version = sdist_reader.get_package_name_and_version(
+                filepath)
+        assert name == 'foo-bar'
+        assert version == '1.0-2b'
+
+    def test_both_zip(self, osutils, sdist_reader):
+        setup_py = self._SETUP_PY % (
+            self._BOTH, 'foo', '1.0'
+        )
+        with osutils.tempdir() as tempdir:
+            filepath = self._write_fake_sdist(setup_py, tempdir, 'zip')
+            name, version = sdist_reader.get_package_name_and_version(
+                filepath)
+        assert name == 'foo'
+        assert version == '1.0'
+
+    def test_bad_format(self, osutils, sdist_reader):
+        setup_py = self._SETUP_PY % (
+            self._BOTH, 'foo', '1.0'
+        )
+        with osutils.tempdir() as tempdir:
+            filepath = self._write_fake_sdist(setup_py, tempdir, 'tar.gz2')
+            with pytest.raises(InvalidSourceDistributionNameError):
+                name, version = sdist_reader.get_package_name_and_version(
+                    filepath)
+
+
+class TestPackage(object):
+
+    def test_same_pkg_sdist_and_wheel_collide(self, osutils, sdist_builder):
+        with osutils.tempdir() as tempdir:
+            sdist_builder.write_fake_sdist(tempdir, 'foobar', '1.0')
+            pkgs = set()
+            pkgs.add(Package('', 'foobar-1.0-py3-none-any.whl'))
+            pkgs.add(Package(tempdir, 'foobar-1.0.zip'))
+            assert len(pkgs) == 1
